telegram_calculator/backend/requirements.txt
aiogram==3.1.1
aiohttp==3.8.6
python-dotenv==1.0.0